WIP on a sigproc refactor
