+++ plaster/run/sigproc_v2/sigproc_v2_task.py
@@ -15,21 +15,16 @@ class SigprocV2Task(PipelineTask):
 
         ims_import_result = ImsImportResult.load_from_folder(self.inputs.ims_import)
 
-        sigproc_v2_params.set_radiometry_channels_from_input_channels_if_needed(
-            ims_import_result.n_channels
-        )
+        # sigproc_v2_params.set_radiometry_channels_from_input_channels_if_needed(
+        #     ims_import_result.n_channels
+        # )
 
-        if sigproc_v2_params.mode == common.SIGPROC_V2_PSF_CALIB:
-            sigproc_v2_instrument_calib_result = worker.sigproc_psf_calib(
+        if sigproc_v2_params.mode in (common.SIGPROC_V2_PSF_CALIB, common.SIGPROC_V2_PSF_CALIB):
+            sigproc_v2_instrument_calib_result = worker.sigproc_instrument_calib(
                 sigproc_v2_params, ims_import_result, self.progress
             )
             sigproc_v2_instrument_calib_result.save()
-
-        if sigproc_v2_params.mode == common.SIGPROC_V2_ILLUM_CALIB:
-            sigproc_v2_instrument_calib_result = worker.sigproc_illum_calib(
-                sigproc_v2_params, ims_import_result, self.progress
-            )
-            sigproc_v2_instrument_calib_result.save()
+            sigproc_v2_instrument_calib_result.calib.save(sigproc_v2_params.calibration_file)
 
         elif sigproc_v2_params.mode == common.SIGPROC_V2_INSTRUMENT_ANALYZE:
             sigproc_v2_result = worker.sigproc_analyze(
